+++ custom_components/callmonitor_test/__init__.py
@@ -8,7 +8,7 @@ from homeassistant.components.http import StaticPathConfig
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant, ServiceCall
 
-from .const import DOMAIN, PLATFORMS, SERVICE_ADD_CONTACT, SERVICE_CLEAR_CALLS, SERVICE_DELETE_CALL, SERVICE_SYNC_PHONEBOOK, STATIC_URL
+from .const import DOMAIN, PLATFORMS, SERVICE_ADD_CONTACT, SERVICE_CLEAR_CALLS, SERVICE_DELETE_CALL, SERVICE_SYNC_ANSWERING_MACHINE, SERVICE_SYNC_PHONEBOOK, STATIC_URL
 
 
 async def async_setup_entry(
@@ -43,6 +43,12 @@ async def async_setup_entry(
         if sensor is not None:
             await sensor.async_sync_phonebook()
 
+    async def async_sync_answering_machine(call: ServiceCall) -> None:
+        """Synchronize FRITZ!Box voicemail messages."""
+        sensor = hass.data[DOMAIN].get(entry.entry_id)
+        if sensor is not None:
+            await sensor.async_sync_answering_machine()
+
     async def async_add_contact(call: ServiceCall) -> None:
         """Add a contact to a FRITZ!Box phonebook."""
         sensor = hass.data[DOMAIN].get(entry.entry_id)
@@ -71,6 +77,13 @@ async def async_setup_entry(
             DOMAIN,
             SERVICE_SYNC_PHONEBOOK,
             async_sync_phonebook,
+        )
+
+    if not hass.services.has_service(DOMAIN, SERVICE_SYNC_ANSWERING_MACHINE):
+        hass.services.async_register(
+            DOMAIN,
+            SERVICE_SYNC_ANSWERING_MACHINE,
+            async_sync_answering_machine,
         )
 
     if not hass.services.has_service(DOMAIN, SERVICE_ADD_CONTACT):
@@ -108,6 +121,8 @@ async def async_unload_entry(
             hass.services.async_remove(DOMAIN, SERVICE_CLEAR_CALLS)
         if hass.services.has_service(DOMAIN, SERVICE_SYNC_PHONEBOOK):
             hass.services.async_remove(DOMAIN, SERVICE_SYNC_PHONEBOOK)
+        if hass.services.has_service(DOMAIN, SERVICE_SYNC_ANSWERING_MACHINE):
+            hass.services.async_remove(DOMAIN, SERVICE_SYNC_ANSWERING_MACHINE)
         if hass.services.has_service(DOMAIN, SERVICE_ADD_CONTACT):
             hass.services.async_remove(DOMAIN, SERVICE_ADD_CONTACT)
         if hass.services.has_service(DOMAIN, SERVICE_DELETE_CALL):